+++ src/file-utils/tmp-file-writer.ts
@@ -15,32 +15,58 @@ import {
 } from './create-tmp-file'
 
 export interface TmpFileWriterOptions extends CreateTmpFileOptions {
+  stopWriteWhenClosed?: boolean
+  enableProcessLoading?: boolean
+  autoSaveWhenDone?: boolean
+  autoCloseWhenDone?: boolean
   buildAiStream: () => Promise<IterableReadableStream<AIMessageChunk>>
   onCancel?: () => void
 }
 
+/**
+ * Writes temporary file with AI-generated content.
+ *
+ * @param options - The options for writing the temporary file.
+ * @returns A promise that resolves to the result of writing the temporary file.
+ */
 export const tmpFileWriter = async (
   options: TmpFileWriterOptions
 ): Promise<WriteTmpFileResult> => {
-  const { buildAiStream, onCancel, ...createTmpFileOptions } = options
+  const {
+    buildAiStream,
+    onCancel,
+    stopWriteWhenClosed = true,
+    enableProcessLoading = true,
+    autoSaveWhenDone = false,
+    autoCloseWhenDone = false,
+    ...createTmpFileOptions
+  } = options
 
   const createTmpFileAndWriterReturns =
     await createTmpFileAndWriter(createTmpFileOptions)
-  const { writeTextPart, getText, writeText, isClosedWithoutSaving } =
-    createTmpFileAndWriterReturns
+  const {
+    writeTextPart,
+    getText,
+    writeText,
+    save,
+    close,
+    isClosedWithoutSaving
+  } = createTmpFileAndWriterReturns
 
   const ModelProvider = await getCurrentModelProvider()
   const { showProcessLoading, hideProcessLoading } = createLoading()
 
   try {
-    showProcessLoading({
-      onCancel
-    })
+    enableProcessLoading &&
+      showProcessLoading({
+        onCancel
+      })
+
     const aiStream = await buildAiStream()
 
     for await (const chunk of aiStream) {
-      if (isClosedWithoutSaving()) {
-        hideProcessLoading()
+      if (stopWriteWhenClosed && isClosedWithoutSaving()) {
+        enableProcessLoading && hideProcessLoading()
         return createTmpFileAndWriterReturns
       }
 
@@ -50,33 +76,32 @@ export const tmpFileWriter = async (
 
       // remove code block syntax
       // for example, remove ```python\n and \n```
-      const currentCode = getText()
-      const removeCodeBlockStartSyntaxCode =
-        removeCodeBlockStartSyntax(currentCode)
+      const currentText = getText()
+      const cleanedText = removeCodeBlockStartSyntax(currentText)
 
-      if (removeCodeBlockStartSyntaxCode !== currentCode) {
-        await writeText(removeCodeBlockStartSyntaxCode)
+      if (cleanedText !== currentText) {
+        await writeText(cleanedText)
       }
-    }
-
-    // remove code block end syntax
-    // for example, remove \n``` at the end
-    const currentCode = getText()
-    const removeCodeBlockEndSyntaxCode = removeCodeBlockEndSyntax(currentCode)
-
-    if (removeCodeBlockEndSyntaxCode !== currentCode) {
-      await writeText(removeCodeBlockEndSyntaxCode)
     }
 
     // remove code block syntax
     // for example, remove ```python\n and \n``` at the start and end
     // just confirm the code is clean
-    const finalCode = removeCodeBlockSyntax(getText())
+    const currentText = getText()
+    const finalText = removeCodeBlockSyntax(
+      removeCodeBlockEndSyntax(currentText)
+    )
 
-    // write the final code
-    await writeText(finalCode)
+    if (finalText !== currentText) {
+      // write the final code
+      await writeText(finalText)
+    }
+
+    if (autoSaveWhenDone) save()
+
+    if (autoCloseWhenDone) close()
   } finally {
-    hideProcessLoading()
+    enableProcessLoading && hideProcessLoading()
   }
 
   return createTmpFileAndWriterReturns