+++ website/.vitepress/theme/index.ts
@@ -1,6 +1,11 @@
+import { NolebaseInlineLinkPreviewPlugin } from '@nolebase/vitepress-plugin-inline-link-preview/client'
 import DefaultTheme from 'vitepress/theme'
 import type { App } from 'vue'
 
+import '@nolebase/vitepress-plugin-inline-link-preview/client/style.css'
+
+import AideModelPrice from '../../components/AideModels/AideModelPrice/index.vue'
+import AidePay from '../../components/AideModels/AidePay.vue'
 import Video from '../../components/Video.vue'
 
 import 'virtual:uno.css'
@@ -10,5 +15,8 @@ export default {
   extends: DefaultTheme,
   enhanceApp({ app }: { app: App }) {
     app.component('Video', Video)
+    app.component('AidePay', AidePay)
+    app.component('AideModelPrice', AideModelPrice)
+    app.use(NolebaseInlineLinkPreviewPlugin)
   }
 }